requirements.txt
openai>=1.0
sounddevice
numpy
pynput
pystray
Pillow
pyperclip
deepgram-sdk>=3.0,<4.0
# faster-whisper>=1.0  # optional: required only for "local" STT provider
